# docs/devicerepository_swagger.yaml
basePath: /
definitions:
...
  models.Attribute:
    properties:
      key:
        type: string
      origin:
        type: string
      value:
        type: string
    type: object
...
  models.Content:
    properties:
      content_variable:
        $ref: '#/definitions/models.ContentVariable'
      id:
        type: string
      protocol_segment_id:
        type: string
      serialization:
        $ref: '#/definitions/models.Serialization'
    type: object
  models.ContentVariable:
    properties:
      aspect_id:
        type: string
      characteristic_id:
        type: string
      function_id:
        type: string
      id:
        type: string
      is_void:
        type: boolean
      name:
        type: string
      omit_empty:
        type: boolean
      serialization_options:
        items:
          type: string
        type: array
      sub_content_variables:
        items:
          $ref: '#/definitions/models.ContentVariable'
        type: array
      type:
        $ref: '#/definitions/models.Type'
      unit_reference:
        type: string
      value: {}
    type: object
...
  models.DeviceType:
    properties:
      attributes:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      description:
        type: string
      device_class_id:
        type: string
      id:
        type: string
      name:
        type: string
      service_groups:
        items:
          $ref: '#/definitions/models.ServiceGroup'
        type: array
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
    type: object
...
  models.Interaction:
    enum:
    - event
    - request
    - event+request
    type: string
    x-enum-varnames:
    - EVENT
    - REQUEST
    - EVENT_AND_REQUEST
...
  models.Serialization:
    enum:
    - xml
    - json
    - plain-text
    type: string
    x-enum-varnames:
    - XML
    - JSON
    - PlainText
  models.Service:
    properties:
      attributes:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      description:
        type: string
      id:
        type: string
      inputs:
        items:
          $ref: '#/definitions/models.Content'
        type: array
      interaction:
        $ref: '#/definitions/models.Interaction'
      local_id:
        type: string
      name:
        type: string
      outputs:
        items:
          $ref: '#/definitions/models.Content'
        type: array
      protocol_id:
        type: string
      service_group_key:
        type: string
    type: object
  models.ServiceGroup:
    properties:
      description:
        type: string
      key:
        type: string
      name:
        type: string
    type: object
  models.Type:
    enum:
    - https://schema.org/Text
    - https://schema.org/Integer
    - https://schema.org/Float
    - https://schema.org/Boolean
    - https://schema.org/ItemList
    - https://schema.org/StructuredValue
    type: string
    x-enum-varnames:
    - String
    - Integer
    - Float
    - Boolean
    - List
    - Structure
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Repository API
  version: "0.1"
paths:
...
  /device-types:
    post:
      description: create device-type
      parameters:
      - description: comma separated list of attribute keys; no other device-type
          with the same attribute key/value may exist
        in: query
        name: distinct_attributes
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DeviceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceType'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create device-type
      tags:
      - device-types
    put:
      consumes:
      - application/json
      description: validate device-type
      parameters:
      - description: must be true; reminder, that this is not an update but a validation
        in: query
        name: dry-run
        required: true
        type: boolean
      - description: Device-Type to be validated
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DeviceType'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: validate device-type
      tags:
      - device-types
  /device-types/{id}:
    delete:
      description: delete device-type
      parameters:
      - description: DeviceType Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete device-type
      tags:
      - device-types
    get:
      description: get device-type
      parameters:
      - description: Device-Type Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceType'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get device-type
      tags:
      - device-types
    put:
      description: set device-type
      parameters:
      - description: DeviceType Id
        in: path
        name: id
        required: true
        type: string
      - description: comma separated list of attribute keys; no other device-type
          with the same attribute key/value may exist
        in: query
        name: distinct_attributes
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DeviceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceType'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device-type
      tags:
      - device-types
...
  /v3/device-types:
    get:
      description: list device-types
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter
        in: query
        name: search
        type: string
      - description: default name.asc
        in: query
        name: sort
        type: string
      - description: filter; ignores limit/offset; comma-separated list
        in: query
        name: ids
        type: string
      - description: filter; comma-separated list; lists elements only if they use
          a protocol that is in the given list
        in: query
        name: protocol-ids
        type: string
      - description: filter; comma-separated list; lists elements only if they have
          an attribute key that is in the given list
        in: query
        name: attr-keys
        type: string
      - description: filter; comma-separated list; lists elements only if they have
          an attribute value that is in the given list
        in: query
        name: attr-values
        type: string
      - description: include id-modified device-types
        in: query
        name: include-modified
        type: boolean
      - description: no unmodified device-types
        in: query
        name: ignore-unmodified
        type: boolean
      - description: filter; json encoded []model.FilterCriteria
        in: query
        name: criteria
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeviceType'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list device-types
      tags:
      - device-types
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"